+++ internal/handler/job/create_test.go
@@ -11,6 +11,7 @@ package job
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/siemens/wfx/generated/api"
 	"github.com/siemens/wfx/internal/handler/job/events"
@@ -39,7 +40,7 @@ func TestCreateJob_Notification(t *testing.T) {
 	db := newInMemoryDB(t)
 	wf := createDirectWorkflow(t, db)
 
-	ch := events.AddSubscriber(context.Background(), events.FilterParams{}, nil)
+	subscriber := events.AddSubscriber(t.Context(), time.Minute, events.FilterParams{}, nil)
 
 	job, err := CreateJob(context.Background(), db, &api.JobRequest{
 		ClientID: "foo",
@@ -47,7 +48,7 @@ func TestCreateJob_Notification(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	jobEvent := <-ch
+	jobEvent := <-subscriber.Events
 	assert.Equal(t, events.ActionCreate, jobEvent.Action)
 	assert.Equal(t, job.ID, jobEvent.Job.ID)
 }