+++ internal/handler/job/delete_test.go
@@ -12,6 +12,7 @@ import (
 	"context"
 	"errors"
 	"testing"
+	"time"
 
 	"github.com/Southclaws/fault/ftag"
 	"github.com/siemens/wfx/generated/api"
@@ -29,13 +30,13 @@ func TestDeleteJob(t *testing.T) {
 	job, err := db.CreateJob(context.Background(), &tmpJob)
 	require.NoError(t, err)
 
-	ch := events.AddSubscriber(context.Background(), events.FilterParams{}, nil)
+	subscriber := events.AddSubscriber(t.Context(), time.Minute, events.FilterParams{}, nil)
 	require.NoError(t, err)
 
 	err = DeleteJob(context.Background(), db, job.ID)
 	require.NoError(t, err)
 
-	jobEvent := <-ch
+	jobEvent := <-subscriber.Events
 	assert.Equal(t, events.ActionDelete, jobEvent.Action)
 	assert.Equal(t, job.ID, jobEvent.Job.ID)
 }